# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Printer Ink Inventory Management API – Elite Fire Protection Systems
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          description: Invalid input or username taken
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found

  # Employees
...
components:
  schemas:
...
    ManagedUser:
      type: object
      required: [id, username, fullName, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, staff]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, fullName, role, password]
      properties:
        username:
          type: string
          minLength: 1
        fullName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, staff]
        password:
          type: string
          minLength: 6
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        fullName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, staff]
        password:
          type: string
          minLength: 6

    # Employee schemas
...