# lib/api-zod/src/generated/types/managedUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Printer Ink Inventory Management API – Elite Fire Protection Systems
 * OpenAPI spec version: 0.1.0
 */
import type { ManagedUserRole } from './managedUserRole';

export interface ManagedUser {
  id: number;
  username: string;
  fullName: string;
  role: ManagedUserRole;
  createdAt: string;
}