# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Printer Ink Inventory Management API – Elite Fire Protection Systems
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  /** @minLength 1 */
  username?: string;
  /** @minLength 1 */
  fullName?: string;
  role?: UserUpdateRole;
  /** @minLength 6 */
  password?: string;
}